feat(go.mod, go.sum): update dependencies
This commit updates the dependencies in `go.mod` and `go.sum`
to their latest versions. This includes updates to libraries such
as `github.com/attestantio/go-eth2-client`,
`github.com/ethpandaops/beacon`, and `github.com/ethpandaops/xatu`.

The reason for this update is to ensure that the project is using
the latest versions of its dependencies, which may include bug
fixes, performance improvements, and new features. Keeping
dependencies up-to-date is a good practice for maintaining the
stability and security of the project.